Add files via upload
This project is a basic Temperature Converter built using HTML, CSS, and JavaScript.
The user can type in a temperature value and choose whether they want to convert from Celsius to Fahrenheit or Fahrenheit to Celsius.

When the user clicks the "Submit" button, the JavaScript function checks whether the input is valid and then performs the correct temperature conversion based on the selected option. The converted value is shown below the button with two decimal accuracy.

This project is beginner-friendly and helps in practicing DOM manipulation, input validation, and simple mathematical operations in JavaScript.